Chapter2:  - Moved classes belonging to Chapter 2 inside the package gettingstarted, similar to the repo fpinscala original from the book
