+++ src/main scala/functional.programming/gettingstarted/CurryOperator.scala
@@ -1,4 +1,4 @@
-package functional.programming
+package functional.programming.gettingstarted
 
 /**
  * Exercise 2.3, 2.4, 2.5
