+++ src/main/scala/functional.programming/gettingstarted/ArrayUtils.scala
@@ -1,4 +1,4 @@
-package functional.programming
+package functional.programming.gettingstarted
 
 import scala.annotation.tailrec
