+++ src/test/scala/functional.programming/gettingstarted/FibonacciCalculatorSpec.scala
@@ -1,4 +1,4 @@
-package functional.programming
+package functional.programming.gettingstarted
 
 import org.scalatest.{Matchers, WordSpec}
